[OBMIN-363] deleted component moveSun
