--- src/components/search-result/moveSun/index.tsx
@@ -1,6 +1,0 @@
-import * as Styles from './styles';
-import { IWindowWidth } from '../types';
-
-export function MoveSun({ windowWidth }: IWindowWidth) {
-  return <Styles.DefaultSun windowWidth={windowWidth} />;
-}
